Major refactoring exercise (Recommender Systems).
The new packaging will disentangle the code from the content of the book. The iweb2.* packages will remain intact to support the audience of the book. The code in the new packages, wherever it overlaps with the book material, will be close to the original but there will be no guarantees of that.

The main idea is to structure the code of the Yooreeka project in a way that can grow fast and allow easy maintenance.